custom_img_gen/requirements.txt
google-cloud-aiplatform
streamlit
Pillow
numpy
pyyaml
requests